create NewsCategoryScreen
