+++ news/list/src/main/java/com/gradlevv/list/ui/NewsCategoryNavigation.kt
@@ -1,0 +1,20 @@
+package com.gradlevv.list.ui
+
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.composable
+import kotlinx.serialization.Serializable
+
+fun NavGraphBuilder.categoryScreen(navHostController: NavHostController) {
+    composable<CategoryDestination> {
+        NewsCategoryScreen(navHostController)
+    }
+}
+
+@Serializable
+data object CategoryDestination
+
+fun NavController.navigateToCategoryScreen() {
+    navigate(CategoryDestination)
+}
